pyproject.toml
[tool.poetry]
name = "gitarchive"
version = "0.1.0"
description = "Archive git repos"
authors = ["Joakim Hellsén <tlovinator@gmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
python-multipart = "^0.0.5"
pytest = "^7.1.2"
Jinja2 = "^3.1.2"
PyGithub = "^1.55"
py-gitea = "^0.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
